task_g_028: Add GetHdyHeader() and SetHdyHeader() to Isolate class
This adds GetHdyHeader() and SetHdyHeader() method to the Isolate
class. It also adds the internal stroage for saving the x-hdy
headers.

Also, create a custom function and private data. For now, it is not
used. It's just in case we need to expand more functionality from this
class. When that happens, we can simply only change the api.cc to
avoid a large dependency chain (large dependency chain happens from
changing a header file).

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>